+++ pages/docs/[[...doc]].tsx
@@ -26,25 +26,12 @@ import { Meta } from '../../components/common/Head/Meta';
 import { HighlightCodeBlock } from '../../components/Docs/HighlightCodeBlock/HighlightCodeBlock';
 import { DOCS_REDIRECTS } from '../../middleware';
 import DocSearchbar from '../../components/Docs/DocSearchbar/DocSearchbar';
+import {
+  IGNORED_DOCS_PATHS,
+  processDocPath,
+  removeOrderingPrefix,
+} from '../api/docs/github';
 
-// ignored files from https://github.com/highlight-run/docs
-export const IGNORED_DOCS_PATHS = new Set<string>([
-  '.git',
-  '.github',
-  '.gitignore',
-  '.husky',
-  '.vscode',
-  '.prettierrc',
-  'node_modules',
-  'package.json',
-  'yarn.lock',
-  'CHANGELOG.md',
-  'CODE_OF_CONDUCT.md',
-  'CONTRIBUTING.md',
-  'LICENSE',
-  'README.md',
-  'SECURITY.md',
-]);
 const DOCS_CONTENT_PATH = path.join(process.cwd(), 'docs');
 
 interface DocPath {
@@ -119,38 +106,6 @@ const useIntersectionObserver = (setActiveId: (s: string) => void) => {
 
 const isValidDirectory = (files: string[]) => {
   return files.find((filename) => filename.includes('index.md')) != null;
-};
-
-const removeOrderingPrefix = (path: string) => {
-  const arrayPath = path.split('/');
-  const cleanPath = arrayPath.map((p) => {
-    const prefixLocation = p.indexOf('_');
-    return prefixLocation === -1 ? p : p.slice(prefixLocation + 1);
-  });
-  return cleanPath.join('/');
-};
-
-export const processDocPath = function (base: string, fileString: string) {
-  const simple_path = path.join(base, fileString);
-  let pp: string;
-  if (fileString.includes('index.md')) {
-    // index.md contains the title of a subheading, which can't have content. get rid of "index.md" at the end
-    pp = simple_path.split('/').slice(0, -1).join('/');
-  } else {
-    // strip out any notion of ".md"
-    pp = simple_path.replace('.md', '');
-    const pp_array = pp.split('/');
-    if (pp_array.length > 1) {
-      const parentDirectory = pp_array[pp_array.length - 2];
-      if (
-        pp_array[pp_array.length - 1] ===
-        `${removeOrderingPrefix(parentDirectory)}-overview`
-      ) {
-        pp = [...pp_array.slice(0, -1), 'overview'].join('/');
-      }
-    }
-  }
-  return removeOrderingPrefix(pp);
 };
 
 // we need to explicitly pass in 'fs_api' because webpack isn't smart enough to
